Рефактор Register, Login Pages
